functions/model-armor/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
google-cloud-modelarmor